Finish Display Host on Android

+++ android/app/src/main/java/com/metallic/chiaki/common/ManualHost.kt
@@ -34,7 +34,7 @@ import io.reactivex.Flowable
 data class ManualHost(
 	@PrimaryKey(autoGenerate = true) val id: Long = 0,
 	val host: String,
-	@ColumnInfo(name = "registered_host") val registeredHost: Int?
+	@ColumnInfo(name = "registered_host") val registeredHost: Long?
 )
 
 @Dao
